Use xxHash for CacheID

+++ src/OpenColorIO/Config.cpp
@@ -4412,7 +4412,7 @@ const char * Config::getCacheID(const ConstContextRcPtr & context) const
         std::ostringstream cacheid;
         serialize(cacheid);
         const std::string fullstr = cacheid.str();
-        getImpl()->m_cacheidnocontext = CacheIDHash(fullstr.c_str(), (int)fullstr.size());
+        getImpl()->m_cacheidnocontext = CacheIDHash(fullstr.c_str(), fullstr.size());
     }
 
     // Also include all file references, using the context (if specified)
@@ -4449,7 +4449,7 @@ const char * Config::getCacheID(const ConstContextRcPtr & context) const
         }
 
         const std::string fullstr = filehash.str();
-        fileReferencesFastHash = CacheIDHash(fullstr.c_str(), (int)fullstr.size());
+        fileReferencesFastHash = CacheIDHash(fullstr.c_str(), fullstr.size());
     }
 
     getImpl()->m_cacheids[contextcacheid] = getImpl()->m_cacheidnocontext + ":" + fileReferencesFastHash;
@@ -5062,4 +5062,3 @@ void Config::archive(std::ostream & ostream) const
 }
 
 } // namespace OCIO_NAMESPACE
-
